Move hash functions to their own module
As with Streams, we move the has functions to their own moduels, and
rename them to be consistent with Redis command names. Their types are
fixed as well, and a corresponding test suite is added.

Along with this, we move some common utility functions to
Cache.Internal. It would be good to find a way to hide this from being
exported.