Properly expose the asynchronous nature of connection creation
redis.creatClient() occurs asynchronously, but we were using the
connection without waiting for it to be available. This resulted in all
commands being queued asynchronously and then being dispatched, which
does not seem to be appropriate.

Int the process, we rename getConn as newConn to correctly reflect that
this creates a new connection.