Convert CacheConn to a typeclass
This adds a SimpleConn type for connections, and changes most of the API
to take an instance of the CacheConn typeclass (which SimpleConn
implements).

The idea is that we can reuse this typeclass for cluster connections in
following commits. We are abusing this method to reuse the Redis command
functions between simple and cluster connections in the JavaScript code.