Make set return a Boolean instead of Unit
Because set can legitimately fail without an error for the NX and EX
cases, we should signal those failures in the return value.

This also clarifies that the return value from JavaScript is not always
a string, but may be null if the NX/EX set fails. We now correctly
handle that.

Added tests for both cases.